AppointmentPolicy for AdminTimeOffController

+++ app/Http/Controllers/AdminTimeOffController.php
@@ -2,12 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use Carbon\Carbon;
-use App\Models\Appointment;
-use App\Models\Barber;
 use App\Models\User;
+use App\Models\Barber;
 use App\Models\Service;
+use App\Models\Appointment;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AdminTimeOffController extends Controller
 {
@@ -123,7 +124,7 @@ class AdminTimeOffController extends Controller
 
     public function show(Appointment $time_off)
     {
-        if ($time_off->service_id !== 1) {
+        if (Gate::denies('isTimeOff',$time_off)) {
             return redirect()->route('bookings.show',$time_off);
         }
 
@@ -135,11 +136,12 @@ class AdminTimeOffController extends Controller
 
     public function edit(Appointment $time_off)
     {
-        if ($time_off->deleted_at) {
-            return redirect()->route('admin-time-offs.show',$time_off)->with('error',"You can't edit cancelled time offs!");
-        } elseif ($time_off->app_start_time <= now()) {
-            return redirect()->route('admin-time-offs.show',$time_off)->with('error',"You can't edit time offs from the past!");
-        } elseif ($time_off->service_id !== 1) {
+        $response = Gate::inspect('adminUpdate',$time_off);
+        if ($response->denied()) {
+            return redirect()->back()->with('error',$response->message());
+        }
+
+        if (Gate::denies('isTimeOff',$time_off)) {
             return redirect()->route('bookings.edit',$time_off);
         }
 
@@ -156,6 +158,15 @@ class AdminTimeOffController extends Controller
 
     public function update(Request $request, Appointment $time_off)
     {
+        $response = Gate::inspect('adminUpdate',$time_off);
+        if ($response->denied()) {
+            return redirect()->back()->with('error',$response->message());
+        }
+
+        if (Gate::denies('isTimeOff',$time_off)) {
+            return redirect()->route('bookings.show',$time_off);
+        }
+
         $request->validate([
             'app_start_date' => ['required','date','after_or_equal:today'],
             'app_start_hour' => ['nullable','integer','between:10,19'],
@@ -183,12 +194,13 @@ class AdminTimeOffController extends Controller
 
     public function destroy(Appointment $time_off)
     {
-        if ($time_off->app_start_time < now()) {
-            return redirect()->back()->with('error',"You can't cancel a previous time off!");
-        } elseif (isset($time_off->deleted_at)) {
-            return redirect()->back()->with('error',"You can't cancel an already cancelled time off!");
-        } elseif ($time_off->service_id !== 1) {
-            return redirect()->route('bookings.show',$time_off)->with('error', "You can't cancel a booking as a time off. Please try again here!");
+        $response = Gate::inspect('adminDelete',$time_off);
+        if ($response->denied()) {
+            return redirect()->back()->with('error',$response->message());
+        }
+
+        if (Gate::denies('isTimeOff',$time_off)) {
+            return redirect()->route('bookings.show',$time_off);
         }
 
         $time_off->delete();
